# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Avito User segmentation application",
        "title": "Swagger Avito User segmentation application",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/adding_user_to_segment": {
            "post": {
                "description": "Addition segments for user in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dependencies"
                ],
                "summary": "AddDelSegments",
                "parameters": [
                    {
                        "description": "Adding Segment",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.dependenciesData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/create_segment": {
            "post": {
                "description": "Create segment in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "CreateSegment",
                "parameters": [
                    {
                        "description": "Create Segment",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.createSegment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/create_user": {
            "post": {
                "description": "Create user in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "CreateUser",
                "parameters": [
                    {
                        "description": "Create User",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/deleting_segment": {
            "post": {
                "description": "Delet segment in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "segments"
                ],
                "summary": "DeletingSegment",
                "parameters": [
                    {
                        "description": "Delete Segment",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.deleteSegment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/deleting_user": {
            "post": {
                "description": "Delete user in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "DeletingUser",
                "parameters": [
                    {
                        "description": "Delete User",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/getting_active_user_segments": {
            "post": {
                "description": "User segment check",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dependencies"
                ],
                "summary": "GettingActiveUserSegments",
                "parameters": [
                    {
                        "description": "Check Segment",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.user"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/history": {
            "post": {
                "description": "Check history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "Hishtory",
                "parameters": [
                    {
                        "description": "ttl hadler",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.history"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        },
        "/ttl_adding_user_to_segment": {
            "post": {
                "description": "Ttl adding or remove",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ttl"
                ],
                "summary": "TtlAddDelSegments",
                "parameters": [
                    {
                        "description": "ttl hadler",
                        "name": "input",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handler.ttlStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createSegment": {
            "type": "object",
            "properties": {
                "percent": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "handler.deleteSegment": {
            "type": "object",
            "properties": {
                "slug": {
                    "type": "string"
                }
            }
        },
        "handler.dependenciesData": {
            "type": "object",
            "properties": {
                "add_segments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "del_segments": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "handler.history": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                }
            }
        },
        "handler.ttlStruct": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.dependenciesData"
                },
                "start": {
                    "type": "string"
                },
                "stop": {
                    "type": "string"
                }
            }
        },
        "handler.user": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "response.HttpError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.HttpResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}